airt/requirements.txt
airtest==1.2.3
anyio==3.4.0
appdirs==1.4.4
attrs==21.2.0
Automat==20.2.0
black==21.7b0
blue==0.7.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
constantly==15.1.0
coverage==6.2
cryptography==36.0.0
cssselect==1.1.0
cycler==0.11.0
ddddocr==1.1.0
decorator==5.1.0
Deprecated==1.2.13
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
easyocr==1.4.1
facebook-wda==1.4.3
flake8==3.8.4
flatbuffers==2.0
fonttools==4.28.2
gitdb==4.0.9
GitPython==3.1.24
green==3.3.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hrpc==1.0.9
httpcore==0.14.3
httpx==0.21.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
imageio==2.9.0
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==2.6.0
kiwisolver==1.3.2
loguru==0.5.3
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.5.0
mccabe==0.6.1
mss==4.0.3
mypy-extensions==0.4.3
networkx==2.6.3
nose==1.3.7
numpy==1.19.3
onnxruntime==1.9.0
opencv-contrib-python==4.5.4.60
opencv-python==3.4.9.31
opencv-python-headless==4.5.4.60
packaging==21.3
parsel==1.6.0
pathspec==0.9.0
Pillow==8.2.0
poco==0.97.0
pocoui==1.0.85
priority==1.3.0
Protego==0.1.16
protobuf==3.19.1
py==1.11.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
PyDispatcher==2.0.5
pyflakes==2.2.0
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-gitlab==1.15.0
python-xlib==0.31
PyWavelets==1.2.0
pywinauto==0.6.3
PyYAML==3.13
queuelib==1.6.2
red==0.2.2
redis==4.0.2
regex==2021.11.10
requests==2.20.1
retry==0.9.2
rfc3986==1.5.0
scikit-image==0.18.3
scipy==1.7.3
Scrapy==2.5.1
service-identity==21.1.0
setuptools-scm==6.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
svn==1.0.1
texttable==0.9.1
tifffile==2021.11.2
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
Twisted==21.7.0
typing_extensions==4.0.0
Unidecode==1.3.2
urllib3==1.24.3
uvloop==0.16.0
w3lib==1.22.0
websocket-client==0.48.0
wrapt==1.13.3
zope.interface==5.4.0
